backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
pydantic[email]==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Security & Rate Limiting
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==42.0.2
pyotp==2.9.0
qrcode[pil]==7.4.2
webauthn==2.0.0

# PostgreSQL
psycopg2-binary==2.9.9

# Monitoring & Error Tracking
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
